requirements.txt
transformers==4.45.1
torch==2.4.1
datasets==3.0.1
scikit-learn==1.5.2
python-dotenv==1.0.1